divide image filters and morphological filters. Class implementation of morphological filters.
